requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.28.0
pydantic>=2.0
pyyaml>=6.0
open_clip_torch>=2.20.0
imageio[ffmpeg]
numpy>=1.23.0,<2
torch>=2.1.0
Pillow>=9.0
